Fix some typings in Tree(Branch) + show schema name as schema tree root node

+++ libs/data-mapper/src/lib/components/canvasToolbox/CanvasToolbox.tsx
@@ -1,5 +1,6 @@
 import { addSourceSchemaNodes, removeSourceSchemaNodes, setCanvasToolboxTabToDisplay } from '../../core/state/DataMapSlice';
 import type { AppDispatch, RootState } from '../../core/state/Store';
+import { SchemaNodeDataType } from '../../models';
 import type { SchemaNodeExtended } from '../../models';
 import { searchSchemaTreeFromRoot } from '../../utils/Schema.Utils';
 import type { ButtonPivotProps } from '../buttonPivot/ButtonPivot';
@@ -7,6 +8,7 @@ import { ButtonPivot } from '../buttonPivot/ButtonPivot';
 import { FloatingPanel } from '../floatingPanel/FloatingPanel';
 import type { FloatingPanelProps } from '../floatingPanel/FloatingPanel';
 import { FunctionList } from '../functionList/FunctionList';
+import { schemaRootKey } from '../targetSchemaPane/TargetSchemaPane';
 import SourceSchemaTreeItem, { useSchemaTreeItemStyles } from '../tree/SourceSchemaTreeItem';
 import Tree from '../tree/Tree';
 import type { ITreeNode } from '../tree/Tree';
@@ -91,6 +93,11 @@ export const CanvasToolbox = ({ canvasBlockHeight }: CanvasToolboxProps) => {
   );
 
   const onSourceSchemaItemClick = (selectedNode: SchemaNodeExtended) => {
+    // If click schema name, just return (don't do anything)
+    if (selectedNode.key === schemaRootKey) {
+      return;
+    }
+
     if (currentSourceSchemaNodes.some((node) => node.key === selectedNode.key)) {
       dispatch(removeSourceSchemaNodes([selectedNode]));
     } else {
@@ -98,16 +105,30 @@ export const CanvasToolbox = ({ canvasBlockHeight }: CanvasToolboxProps) => {
     }
   };
 
-  const searchedSourceSchemaTreeRoot = useMemo<ITreeNode<ITreeNode<SchemaNodeExtended>> | undefined>(() => {
+  const searchedSourceSchemaTreeRoot = useMemo<ITreeNode<SchemaNodeExtended> | undefined>(() => {
     if (!sourceSchema) {
       return undefined;
     }
 
-    if (!sourceSchemaSearchTerm) {
-      return { ...sourceSchema.schemaTreeRoot };
-    } else {
-      return searchSchemaTreeFromRoot(sourceSchema.schemaTreeRoot, sourceSchemaSearchTerm);
+    let newSourceSchemaTreeRoot: ITreeNode<SchemaNodeExtended> = { ...sourceSchema.schemaTreeRoot };
+
+    if (sourceSchemaSearchTerm) {
+      newSourceSchemaTreeRoot = searchSchemaTreeFromRoot(sourceSchema.schemaTreeRoot, sourceSchemaSearchTerm);
     }
+
+    // Format extra top layers to show schema name and schemaTreeRoot
+    // Can safely typecast with the root node(s) as we only use the properties defined here
+    const schemaRoot = {} as ITreeNode<SchemaNodeExtended>;
+    const schemaNameRoot = {} as ITreeNode<SchemaNodeExtended>;
+
+    schemaNameRoot.key = schemaRootKey;
+    schemaNameRoot.name = sourceSchema.name;
+    schemaNameRoot.schemaNodeDataType = SchemaNodeDataType.None;
+    schemaNameRoot.children = [newSourceSchemaTreeRoot];
+
+    schemaRoot.children = [schemaNameRoot];
+
+    return schemaRoot;
   }, [sourceSchema, sourceSchemaSearchTerm]);
 
   const toolboxButtonPivotProps: ButtonPivotProps = useMemo(
@@ -148,15 +169,15 @@ export const CanvasToolbox = ({ canvasBlockHeight }: CanvasToolboxProps) => {
           <Tree<SchemaNodeExtended>
             // Add one extra root layer so schemaTreeRoot is shown as well
             // Can safely typecast as only the children[] are used from root
-            treeRoot={{ children: [searchedSourceSchemaTreeRoot] } as SchemaNodeExtended}
-            nodeContent={(node: SchemaNodeExtended, isHovered: boolean) => (
+            treeRoot={searchedSourceSchemaTreeRoot}
+            nodeContent={(node, isHovered) => (
               <SourceSchemaTreeItem
-                node={node}
+                node={node as SchemaNodeExtended}
                 isNodeAdded={currentSourceSchemaNodes.some((srcSchemaNode) => srcSchemaNode.key === node.key)}
                 isNodeHovered={isHovered}
               />
             )}
-            onClickItem={(node) => onSourceSchemaItemClick(node)}
+            onClickItem={(node) => onSourceSchemaItemClick(node as SchemaNodeExtended)}
             nodeContainerClassName={mergeClasses(schemaNodeItemStyles.nodeContainer, schemaNodeItemStyles.sourceSchemaNode)}
             nodeContainerStyle={(node) => ({
               backgroundColor: currentSourceSchemaNodes.some((srcSchemaNode) => srcSchemaNode.key === node.key)
